preserve and upcast, don't check and downcast

+++ contracts/Interfaces/Oracles/IXRETHOracle.sol
@@ -5,6 +5,6 @@ pragma solidity 0.8.17;
 interface IXRETHOracle {
     /// @dev Oracle data verified using cryptographic fraud proofs.
     /// @return The total value locked in the protocol, in wei, for each minipool + rewards earned.
-    function getTotalYieldAccrued() external view returns (uint);
+    function getTotalYieldAccrued() external view returns (int256);
 
 }
